+++ src/server/services/validation.service.js
@@ -6,7 +6,7 @@ const errorMessages = {
   declaration1: "You must tick all the declarations before continuing",
   declaration2: "You must tick all the declarations before continuing",
   declaration3: "You must tick all the declarations before continuing",
-  operator_type: "You must select an operator type",
+  registration_role: "You must select a role before continuing",
   operator_first_name: "Not a valid first name",
   operator_last_name: "Not a valid last name",
   operator_primary_number: "Not a valid phone number",